requirements.txt
django
google-cloud-texttospeech
pyttsx3
pyaudio
speechrecognition
google-generativeai
gunicorn
whitenoise
dj-database-url
